+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Ensure you use consistent title format.
 - Pin minor version of Python dependency in debian packages.
 - Remove dependency on distutils.
 - Disable 3DES and other loose SSL algorithmes.
+- Supports PostgreSQL 16.
 
 
 **UI changes**